requirements.txt
streamlit==1.41.1
google-generativeai==0.8.4
python-dotenv==1.0.1
pandas==2.2.3
chromadb==0.6.3
beautifulsoup4==4.12.3
requests==2.32.3
Pillow==11.1.0
